+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Select the name automatically when renaming files.
+
 ## 0.2.0 (2019-11-24)
 
 * Close completion popup on saving.